alzheimers_api/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
joblib==1.4.2
numpy==1.26.4
pandas==2.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.1.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
